Added readiness check to avoid errors during topology changes
erpc and nodes show up faster than the applications can start in some
cases leading to requests being routed to booting nodes. A workaround is
possible in the RPC call, but it leads to the proliferation of error
handling code, so it would be better to wait for the nodes to be
ready.